refactor/package/product search package add
