+++ docker-compose.yaml
@@ -9,20 +9,20 @@ services:
     ports:
       - 5000:5000
 
-#  starbucks_database:
-#    platform: linux/x86_64
-#    image: mysql:latest
-#    environment:
-#      MYSQL_DATABASE: 'ghbt'
-#      MYSQL_ROOT_PASSWORD: '1q2w3e4r'
-#    command:
-#      - --character-set-server=utf8mb4
-#      - --collation-server=utf8mb4_unicode_ci
-#      - --skip-character-set-client-handshake
-#    cap_add:
-#      - SYS_NICE
-#    ports:
-#      - 3306:3306
+  starbucks_database:
+    platform: linux/x86_64
+    image: mysql:latest
+    environment:
+      MYSQL_DATABASE: 'ghbt'
+      MYSQL_ROOT_PASSWORD: '1q2w3e4r'
+    command:
+      - --character-set-server=utf8mb4
+      - --collation-server=utf8mb4_unicode_ci
+      - --skip-character-set-client-handshake
+    cap_add:
+      - SYS_NICE
+    ports:
+      - 3306:3306
 
   starbucks_redis:
     image: redis:latest
