+++ src/main/java/com/ghbt/ghbt_starbucks/api/product_search/ISearchCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.ghbt.ghbt_starbucks.api.product_and_category.repository;
+package com.ghbt.ghbt_starbucks.api.product_search;
 
-import com.ghbt.ghbt_starbucks.api.product_and_category.model.SearchCategory;
+import com.ghbt.ghbt_starbucks.api.product_search.SearchCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ISearchCategoryRepository extends JpaRepository<SearchCategory, Long> {
